requirements.txt
PyQt5>=5.15
opencv-python>=4.5
requests>=2.28
pynput>=1.7
numpy>=1.21
PyMuPDF>=1.23
pyinstaller>=6.3.0
playwright>=1.40.0
python-dotenv>=1.0.0
